refactor: Move CommonInterfaces
