+++ packages/tools/devtools/devtools-view/src/components/data-visualization/CommonInterfaces.ts
@@ -4,7 +4,7 @@
  */
 
 /**
- * Base interface used in passing key of the child node from Record {@link VisauTree}
+ * Base interface used in passing key of the child node from Record {@link VisaulTree}
  *
  * @public
  */
